feat: Unlock the user's secret stash of tables
Adds a shiny new /tables route to reveal the tables linked to the logged-in user. This magic happens through the "listUserTables" function in the items controller.

"listUserTables" fetches the user's roles, then grabs the permissions tied to those roles to discover which tables are fair game.
Returns an array of the user's allowed tables, like a treasure map.